Update cookie handling to avoid iOS 7 policy issue; change to userinfo v3 endpoint

+++ trunk/Examples/OAuth2SampleTouch/OAuth2SampleRootViewControllerTouch.m
@@ -173,7 +173,7 @@ static NSString *const kDailyMotionClientSecretKey = @"DailyMotionClientSecret";
 }
 
 - (BOOL)isGoogleSegmentSelected {
-  int segmentIndex = self.serviceSegments.selectedSegmentIndex;
+  NSInteger segmentIndex = self.serviceSegments.selectedSegmentIndex;
   return (segmentIndex == 0);
 }
